+++ base-components/activityCard/types.ts
@@ -1,0 +1,9 @@
+export type TaskCardProps = {
+  title: string;
+  subtitle?: string;
+  progress?: number;
+  total?: number;
+  showProgress?: boolean;
+  iconType?: 'arrow' | 'download' | 'none';
+  onClick?: () => void;
+};
