+++ components/overView/constants.ts
@@ -14,54 +14,16 @@ import {
   FaCalendarAlt,
 } from 'react-icons/fa';
 import { MdChatBubbleOutline } from 'react-icons/md';
-import { IconType } from 'react-icons';
 import { TfiAlignRight } from 'react-icons/tfi';
+import {
+  Task,
+  TQuickAction,
+  AttendanceItem,
+  MessageData,
+  ChecklistItem,
+  Milestone,
+} from './types';
 
-// TYPES
-export type Task = {
-  title: string;
-  subtitle?: string;
-  progress?: number;
-  total?: number;
-  iconType?: 'arrow' | 'download' | 'none';
-  showProgress?: boolean;
-};
-
-export type TQuickAction = {
-  title: string;
-  icon: IconType;
-  iconColor: string;
-};
-
-export type ChecklistItem = {
-  id: number;
-  title: string;
-  description?: string;
-  icon: IconType;
-  progress: string;
-  status: 'completed' | 'in_progress' | 'pending';
-  color: string;
-};
-
-export interface Milestone {
-  icon: IconType;
-  label: string;
-  progress: number; // 0 to 100
-  gradient: string;
-  stars?: number;
-}
-
-export type AttendanceItem = {
-  label: string;
-  value: number;
-};
-
-export type MessageData = {
-  icon: IconType;
-  message: string;
-};
-
-// TASKS
 export const tasks: Task[] = [
   {
     title: 'Activate 5 features',
